Closes #6. Receiver would get called multiple times when a bluetooth device disconnected. Causing the retries to spin out of control and crash the app. Closes #8. Made back button no longer go to last input mode.
Also did UI tests with espresso for android side. Loading tests fail for the gamepad and customize gamepad tests as there is some weird issue where the views are in the view hierarchy but the parent says that they are not its children. Note the parent does include them in its sortedChildren variables but not in its mChildren variable.
UI testing is nowhere near fully functional and modification of our code to support the Dagger framework would be necessary for code coverage.

Also resolved issue where gamepad modes would not change orientation correctly when the loaded layout said they should. Was resetting the orientation to be free for the rest of the app too early.